Renamed namespace RedisMessaging -> Oriflame.RedisMessaging
