+++ src/Oriflame.RedisMessaging.ReliableDelivery/Properties/AssemblyInfo.cs
@@ -1,0 +1,2 @@
+﻿using System.Runtime.CompilerServices;
+[assembly: InternalsVisibleTo("Oriflame.RedisMessaging.ReliableDelivery.Tests")]
