+++ src/Oriflame.RedisMessaging.ReliableDelivery/Publish/IReliablePublisher.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+
+namespace Oriflame.RedisMessaging.ReliableDelivery.Publish
+{
+    public interface IReliablePublisher
+    {
+        void Publish(string channel, string message);
+
+        Task PublishAsync(string channel, string message);
+    }
+}
