+++ src/Oriflame.RedisMessaging.ReliableDelivery/RedisMessaging.ReliableDelivery.ruleset
