+++ src/Oriflame.RedisMessaging.ReliableDelivery/Subscribe/IMessageHandler.cs
@@ -1,0 +1,11 @@
+﻿using StackExchange.Redis;
+
+namespace Oriflame.RedisMessaging.ReliableDelivery.Subscribe
+{
+    public interface IMessageHandler
+    {
+        RedisChannel Channel { get; }
+
+        void HandleMessage(Message message);
+    }
+}
