+++ src/Oriflame.RedisMessaging.ReliableDelivery/Subscribe/IMessageLoader.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Oriflame.RedisMessaging.ReliableDelivery.Subscribe
+{
+    public interface IMessageLoader
+    {
+        IEnumerable<Message> GetMessages(string channelName, long fromMessageId, long toMessageId = long.MaxValue);
+    }
+}
