+++ src/Oriflame.RedisMessaging.ReliableDelivery/Subscribe/IMessageValidationResult.cs
@@ -1,0 +1,4 @@
+﻿namespace Oriflame.RedisMessaging.ReliableDelivery.Subscribe
+{
+    public interface IMessageValidationResult { }
+}
