--- src/RedisMessaging.ReliableDelivery/Publish/IReliablePublisher.cs
@@ -1,11 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace RedisMessaging.ReliableDelivery.Publish
-{
-    public interface IReliablePublisher
-    {
-        void Publish(string channel, string message);
-
-        Task PublishAsync(string channel, string message);
-    }
-}
