--- src/RedisMessaging.ReliableDelivery/Subscribe/IMessageHandler.cs
@@ -1,11 +1,0 @@
-﻿using StackExchange.Redis;
-
-namespace RedisMessaging.ReliableDelivery.Subscribe
-{
-    public interface IMessageHandler
-    {
-        RedisChannel Channel { get; }
-
-        void HandleMessage(Message message);
-    }
-}
