--- src/RedisMessaging.ReliableDelivery/Subscribe/IMessageLoader.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace RedisMessaging.ReliableDelivery.Subscribe
-{
-    public interface IMessageLoader
-    {
-        IEnumerable<Message> GetMessages(string channelName, long fromMessageId, long toMessageId = long.MaxValue);
-    }
-}
