--- src/RedisMessaging.ReliableDelivery/Subscribe/IMessageValidationResult.cs
@@ -1,4 +1,0 @@
-﻿namespace RedisMessaging.ReliableDelivery.Subscribe
-{
-    public interface IMessageValidationResult { }
-}
