--- src/RedisMessaging.ReliableDelivery/Subscribe/IReliableSubscriber.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace RedisMessaging.ReliableDelivery.Subscribe
-{
-    public interface IReliableSubscriber
-    {
-        void Subscribe(IMessageHandler handler);
-
-        Task SubscribeAsync(IMessageHandler handler);
-
-        void Unsubscribe(string channel);
-
-        Task UnsubscribeAsync(string channel);
-
-        void UnsubscribeAll();
-
-        Task UnsubscribeAllAsync();
-    }
-}
