--- test/RedisMessaging.ReliableDelivery.Tests/RedisMessaging.ReliableDelivery.Tests.csproj
@@ -1,27 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-    <IsPackable>false</IsPackable>
-    <ProjectGuid>{C59EBBCC-62A1-4D63-9C82-D507F66E5DE9}</ProjectGuid>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Divergic.Logging.Xunit" Version="1.1.0" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
-    <PackageReference Include="Moq" Version="4.10.1" />
-    <PackageReference Include="RedisInside-Core" Version="3.3.100" />
-    <PackageReference Include="xunit" Version="2.4.1" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.1">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
-    </PackageReference>
-    <DotNetCliToolReference Include="dotnet-xunit" Version="2.3.1" />
-    <PackageReference Include="Ori.Common.SonarQubeTests" Version="1.1.0" PrivateAssets="All" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\RedisMessaging.ReliableDelivery\RedisMessaging.ReliableDelivery.csproj" />
-  </ItemGroup>
-
-</Project>
